feat: stream权限接口优化升级 #5862
