+++ src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/service/stream/StreamPermissionProjectServiceImpl.kt
@@ -34,8 +34,8 @@ import com.tencent.devops.common.auth.api.pojo.BkAuthGroupAndUserList
 import org.springframework.beans.factory.annotation.Autowired
 
 class StreamPermissionProjectServiceImpl @Autowired constructor(
-    val streamPermissionService: StreamPermissionService
-) : PermissionProjectService{
+    val IStreamPermissionValidateService: IStreamPermissionValidateService
+) : PermissionProjectService {
     override fun getProjectUsers(projectCode: String, group: BkAuthGroup?): List<String> {
         // stream场景下使用不到此接口。占做默认实现
         return emptyList()
@@ -56,14 +56,14 @@ class StreamPermissionProjectServiceImpl @Autowired constructor(
         projectCode: String,
         group: BkAuthGroup?
     ): Boolean {
-        return streamPermissionService.isProjectMember(projectCode, userId).first
+        return IStreamPermissionValidateService.isProjectMember(projectCode, userId).first
     }
 
     override fun checkProjectManager(
         userId: String,
         projectCode: String
     ): Boolean {
-        return streamPermissionService.isProjectMember(projectCode, userId).second
+        return IStreamPermissionValidateService.isProjectMember(projectCode, userId).second
     }
 
     override fun createProjectUser(
